Product function

+++ Assignment/web/ManagerProduct.jsp
@@ -48,6 +48,7 @@
                             <th>Size</th>
                             <th>Price</th>
                             <th>Image</th>
+                            <th>Quanlity</th>
                             <th>Description</th>
                             <th>IDcategory</th>
                         </tr>
@@ -69,29 +70,20 @@
                                 <td>
                                     <img src="./images/${item.img}.jpg">
                                 </td>
+                                 <td>${item.quanlity}</td>
                                 <td>${item.description}</td>
                                 <td>${item.IDcategory}</td>
                                 
                                 <td>
                                     <a href="loadproduct?pid=${item.IDproduct}"  class="edit" data-toggle="modal"><i class="material-icons" data-toggle="tooltip" title="Edit">&#xE254;</i></a>
-                                    <a href="delete?pid=${item.IDproduct}" class="delete" data-toggle="modal"><i class="material-icons" data-toggle="tooltip" title="Delete">&#xE872;</i></a>
+                                    <a href="delete?pid=${item.IDproduct}" class="delete" data-toggle="modal" onclick="return confirm('Are you sure?')"><i class="material-icons" data-toggle="tooltip" title="Delete">&#xE872;</i></a>
+                                    
                                 </td>
                             </tr>
                         </c:forEach>
                     </tbody>
                 </table>
-                <div class="clearfix">
-                    <div class="hint-text">Showing <b>5</b> out of <b>25</b> entries</div>
-                    <ul class="pagination">
-                        <li class="page-item disabled"><a href="#">Previous</a></li>
-                        <li class="page-item"><a href="#" class="page-link">1</a></li>
-                        <li class="page-item"><a href="#" class="page-link">2</a></li>
-                        <li class="page-item active"><a href="#" class="page-link">3</a></li>
-                        <li class="page-item"><a href="#" class="page-link">4</a></li>
-                        <li class="page-item"><a href="#" class="page-link">5</a></li>
-                        <li class="page-item"><a href="#" class="page-link">Next</a></li>
-                    </ul>
-                </div>
+               
             </div>
             
 
@@ -120,20 +112,23 @@
                             </div>
                             <div class="form-group">
                                 <label>Size</label>
-                                <textarea name="size" class="form-control" required></textarea>
+                                <input name="size" type="text" class="form-control" required>
                             </div>
                             <div class="form-group">
                                 <label>Price</label>
                                 <input name="price" type="text" class="form-control" required>
                             </div>
-                            
+                            <div class="form-group">
+                                <label>Quanlity</label>
+                                <input name="quanlity" type="text" class="form-control" required>
+                            </div>
                             <div class="form-group">
                                 <label>Description</label>
                                 <textarea name="description" class="form-control" required></textarea>
                             </div>
                             <div class="form-group">
                                 <label>Status</label>
-                                <textarea name="status" class="form-control" required></textarea>
+                                 <input name="status" type="text" class="form-control" required>
                             </div>
                             <div class="form-group">
                                 <label>IDCategory</label>
